refactor: extract match testing methods
